own_MCP_servers/WebSearch/requirements.txt
fastmcp
duckduckgo_search
requests
beautifulsoup4
sumy
